# Improve OCaml runtimes compared to C