--- a/README.md
+++ b/README.md
@@ -1,11 +1,11 @@
-# Contributions from users after the publication of the paper
+# Contributions from users/readers after the publication of the paper
 
 [![DOI](https://zenodo.org/badge/20165/fangohr/paper-supplement-ocaml-performance.svg)](https://zenodo.org/badge/latestdoi/20165/fangohr/paper-supplement-ocaml-performance) <a href="http://arxiv.org/abs/1601.07392"><img src="https://img.shields.io/badge/preprint-arxiv:1601.07392-lightgrey.svg" alt="arxiv"></a>
 <a href="http://dl.acm.org/citation.cfm?doid=2897676.2897677"><img src="https://img.shields.io/badge/journal-SE4Science-blue.svg" alt="URL"></a>
 
 <br>
 
-Contributions from users received **after** the publication of
+Contributions from received **after** the publication of
 ["Nmag micromagnetic simulation tool - software engineering lessons learned"](http://arxiv.org/abs/1601.07392) by Hans Fangohr, Maximilian Albert and Matteo Franchin. In proceedings of the International Workshop on Software Engineering for Science, at ICSE2016 SE4Science '16, 1-7 (2016)
 
 Journal URL: http://dl.acm.org/citation.cfm?doid=2897676.2897677
@@ -16,8 +16,8 @@ DOI: doi = 10.1145/2897676.2897677
 
 This branch of the repository is a fork of the master branch and contains
 additions and improvements possibly contributed by external parties. While the
-master branch is freezed so that it remains consistent with the original
-publication, this branch is free to evolve and incorporate additional
+master branch is frozen so that it remains consistent with the original
+publication, this branch can evolve and incorporate additional
 observations and suggestions. For example, we could update this branch to
 reflect the state of the latest versions of the OCaml and C/C++ compilers.