--- a/array-sum/.gitignore
+++ b/array-sum/.gitignore
@@ -1,0 +1,2 @@
+array_test
+array_test.s