--- a/array-sum/Makefile
+++ b/array-sum/Makefile
@@ -21,3 +21,4 @@ carray_test.s: carray_test.cc
 
 clean:
 	rm -f *.cmi *.cmo *.cmx *.o array_test carray_test
+	rm -f carray_test.s array_test.s