--- a/multidim-arrays/.gitignore
+++ b/multidim-arrays/.gitignore
@@ -1,0 +1,5 @@
+array_of_bigarrays_test
+array_test
+array_test.s
+bigarray_test
+carray_test