--- a/summation/.gitignore
+++ b/summation/.gitignore
@@ -1,0 +1,2 @@
+array_test
+array_test.s