I was interested in your remarks on OCaml vs C performance for numeric code.

Avoiding Bigarray.Genarray gives an order of magnitude performance improvement on my machine (the type annotations and -unsafe seem to have little effect on this example).

Before:

```
time ./bigarray_test
Sum is 0.595922
Create time 3.09944e-06 s
  Init time 1.19804 s
   Sum time 0.446599 s
1.62user 0.03system 0:01.65elapsed 99%CPU (0avgtext+0avgdata 215836maxresident)k
0inputs+0outputs (0major+53400minor)pagefuls 0swaps
```

After:

```
time ./bigarray_test
Sum is 0.595922
Create time 0.214826 s
  Init time 0.730555 s
   Sum time 0.0553079 s
0.94user 0.06system 0:01.01elapsed 99%CPU (0avgtext+0avgdata 304936maxresident)k
0inputs+0outputs (0major+75924minor)pagefuls 0swaps
```