@tbrk 2016-12-04
Sorry for the delay. I have made the changes that you requested on the new branch (contribs_after_publication) and push -f'ed.