@mfnch 2016-12-04
Hi Timothy,
Thanks for your contribution. I merged it on the contribs_after_publication branch. I'll re-run all the examples and upload the new results later this week.
Cheers,
Matteo